[python] extract converter_class.cpp from python_converter.cpp
Move class/instance-attribute methods of python_converter into
src/python-frontend/converter/converter_class.cpp: make_enum_member_struct_expr,
extract_class_name_from_tag, create_normalized_self_key, clean_attribute_type,
create_member_expression, register_instance_attribute, is_instance_attribute,
copy_instance_attributes, update_instance_from_self, infer_attr_type_from_usage,
get_attributes_from_self, process_forward_reference, get_class_definition.

Promote two cross-TU helpers (build_component, is_enum_class) into a new
internal header converter/converter_internal.h under a python_frontend
namespace; update remaining call sites in python_converter.cpp accordingly.

Pure code movement — no logic change.
